Separate SqlClientPermissionAttribute class from SqlClientPermission.cs
